moved student and teacher store to their own file
